10/06 matin, question/questionnaire ok

+++ src/main/java/model/Question.java
@@ -34,7 +34,7 @@ public class Question implements java.io.Serializable, Comparable {
 	private Set<SujetQuestion> sujetQuestions = new HashSet<SujetQuestion>(0);
 	private Set<Reponse> reponses = new HashSet<Reponse>(0);
 	private Set<ReponseApprenant> reponseApprenants = new HashSet<ReponseApprenant>(0);
-	private double tauxreussite=100;
+	private Integer tauxreussite=100;
 	private Integer nbnotes=0;
 	private Integer nbreussite=0;
 	
@@ -158,11 +158,11 @@ public class Question implements java.io.Serializable, Comparable {
 	}
 
 	@Column(name = "tauxreussite")
-	public double getTauxreussite() {
+	public Integer getTauxreussite() {
 		return this.tauxreussite;
 	}
 
-	public void setTauxreussite(double tauxreussite) {
+	public void setTauxreussite(Integer tauxreussite) {
 			this.tauxreussite = tauxreussite;
 	}
